requirements.txt
asgiref==3.8.1
boto3==1.35.36
botocore==1.35.99
click==8.1.8
colorama==0.4.6
decorator==5.1.1
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.12.0
django-import-export==4.3.6
django-redis==5.4.0
django-rq==3.0.0
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
jmespath==1.0.1
moviepy==2.1.2
numpy==2.2.3
packaging==24.2
pillow==10.4.0
pluggy==1.6.0
proglog==0.1.10
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
rq==2.1.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.5.0
whitenoise==6.9.0
